requirements.txt
streamlit
extra-streamlit-components
pyrebase4